pyproject.toml
[build-system]
requires = ["poetry"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "fristad"
version = "2.2.3"
authors = ["Grahame Bowland <grahame@oreamnos.com.au>"]
description = "Template and share OpenSSH ssh_config(5) files."
readme = "README.md"
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Intended Audience :: System Administrators",
  "Topic :: System :: Networking",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX",
]

[tool.poetry.scripts]
fristad = 'fristad.cli:cli'

[tool.poetry.dependencies]
python = ">=3.9"
colormath = "^3.0.0"
paho-mqtt = "^1.6.1"

[tool.poetry.dev-dependencies]
black = "*"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | tests/.*/setup.py
)/
'''
